server/pyproject.toml
[tool.poetry]
name = "helios-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.5"
anyjson = "==0.3.3"
bleach = "==1.4.1"
boto = "==2.27.0"
celery = "==4.3.0"
dj-database-url = "==0.3.0"
django = "==1.11.20"
django-allauth = "^=0.35.0"
django-picklefield = "==0.3.0"
django-secure = "==1.0.1"
django-ses = "==0.6.0"
django-sslify = "==0.2.7"
django-webtest = ">=1.9"
django-environ = "~=0.4.5"
django-cors-headers = "~3.1.1"
django-debug-toolbar = "==2.1"
gunicorn = "==19.9"
html5lib = "==0.999"
kombu = "==4.4.0"
oauth2client = "==4.1.2"
pillow = "==7.0.0"
psycopg2 = "==2.7.3.2"
pyparsing = "==2.4.5"
python-dateutil = ">=1.5"
requests = "==2.21.0"
validate-email = "==1.3"
webtest = "==2.0.18"
django-redis = "~=4.10.0"
redis = "~=3.3.11"
qrcode = "==6.1"
djangorestframework = "^3.11.0"
markdown = "^3.1.1"
django-filter = "^2.2.0"

[tool.poetry.dev-dependencies]
django-extensions= "^=1.9.9"
Werkzeug= "^=0.14.1"
pytest = ">=5.3.2"
pytest-cov = "~=2.8.1"
pytest-django = "~=3.5.1"
pytest-xdist = "~=1.30.0"
pytest-sugar = "~=0.9.2"
pytest-faker = "~=2.0.0"
pytest-mock = "~=1.11.1"
pytest-pudb = "~=0.7.0"
pytest-factoryboy = "~=2.0.3"
ipdb = "^=0.10.3"
ipython-genutils = "^=0.2.0"
ipython = "^=6.2.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
